# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpsReplay API",
    "version": "0.2.0",
    "description": "Proposed REST contract for API Gateway and Lambda. The terminal gateway WebSocket protocol is in docs/api.md. No implementation exists yet."
  },
  "servers": [
    {
      "url": "https://api.opsreplay.example",
      "description": "Placeholder hosted endpoint. The domain is not chosen and no service is running."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/sessions": {
      "post": {
        "operationId": "startSession",
        "summary": "Start a Challenge attempt in a dedicated environment task",
        "description": "Idempotent by requestId. A repeated request returns the saved session or confirmed launch failure and never launches a second task. After a confirmed capacity rejection, wait retryAfterSeconds and use a new request ID to start again.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "get": {
        "operationId": "listSessions",
        "summary": "List owned Challenge attempts, newest first",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 1000
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Cognito-issued token."
      }
    },
    "responses": {
      "Error": {
        "description": "Typed error without hidden Challenge or review answer data.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "INVALID_REQUEST",
              "UNAUTHENTICATED",
              "ACCESS_DENIED",
              "NOT_FOUND",
              "IDEMPOTENCY_CONFLICT",
              "VERSION_UNAVAILABLE",
              "ACTIVE_SESSION_EXISTS",
              "SESSION_NOT_READY",
              "SESSION_ACTIVE",
              "SESSION_TERMINAL",
              "DEBRIEF_PENDING",
              "CAPACITY_UNAVAILABLE",
              "HINT_NOT_AVAILABLE",
              "NO_HINTS_REMAINING",
              "TURN_IN_PROGRESS",
              "LIMIT_EXCEEDED",
              "PROVIDER_FAILED",
              "INTERNAL_ERROR",
              "TURN_INTERRUPTED"
            ]
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "activeSessionId": {
            "type": "string",
            "format": "uuid"
          },
          "retryAfterSeconds": {
            "type": "integer",
            "minimum": 1,
            "maximum": 3600
          }
        },
        "required": ["code", "message", "requestId"]
      },
...
      "Tier": {
        "type": "string",
        "enum": ["easy", "medium", "hard"]
      },
...
      "StartSessionRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "challengeId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "challengeVersion": {
            "type": "string",
            "minLength": 5,
            "maxLength": 30,
            "pattern": "^(0|[1-9][0-9]*)\\.(0|[1-9][0-9]*)\\.(0|[1-9][0-9]*)$"
          }
        },
        "required": ["requestId", "challengeId", "challengeVersion"]
      },
...
      "ChallengeRef": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "version": {
            "type": "string",
            "minLength": 5,
            "maxLength": 30,
            "pattern": "^(0|[1-9][0-9]*)\\.(0|[1-9][0-9]*)\\.(0|[1-9][0-9]*)$"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "tier": {
            "$ref": "#/components/schemas/Tier"
          },
          "category": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          }
        },
        "required": ["id", "version", "title", "tier", "category"]
      },
      "Attempt": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "kind": {
            "type": "string",
            "enum": ["first", "retry"]
          },
          "number": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000
          }
        },
        "required": ["kind", "number"]
      },
      "SessionStatus": {
        "type": "string",
        "enum": ["provisioning", "ready", "resolved", "failed", "ended", "abandoned", "error"]
      },
      "StatusReason": {
        "enum": [
          "validators_passed",
          "time_limit",
          "learner_ended",
          "heartbeat_missed",
          "environment_exited",
          "start_failed",
          null
        ]
      },
      "Alert": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "summary": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "severity": {
            "type": "string",
            "enum": ["warning", "critical"]
          }
        },
        "required": ["title", "summary", "severity"]
      },
      "MetricDescriptor": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40,
            "pattern": "^[a-z][a-z0-9_]*$"
          },
          "label": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "unit": {
            "type": "string",
            "enum": ["requests_per_second", "percent", "milliseconds", "mebibytes"]
          }
        },
        "required": ["id", "label", "unit"]
      },
      "RecoveryState": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "state": {
            "type": "string",
            "enum": ["failing", "sustaining", "met"]
          },
          "sustainedSeconds": {
            "type": "integer",
            "minimum": 0,
            "maximum": 7200
          },
          "requiredSeconds": {
            "type": "integer",
            "minimum": 1,
            "maximum": 600
          }
        },
        "required": ["state", "sustainedSeconds", "requiredSeconds"]
      },
      "Hint": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "releasedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "text", "releasedAt"]
      },
      "Assistance": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "hintsReleased": {
            "type": "integer",
            "minimum": 0,
            "maximum": 50
          },
          "assistantTurns": {
            "type": "integer",
            "minimum": 0,
            "maximum": 500
          },
          "proposalsRun": {
            "type": "integer",
            "minimum": 0,
            "maximum": 500
          }
        },
        "required": ["hintsReleased", "assistantTurns", "proposalsRun"]
      },
      "SessionView": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "challenge": {
            "$ref": "#/components/schemas/ChallengeRef"
          },
          "attempt": {
            "$ref": "#/components/schemas/Attempt"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "statusReason": {
            "$ref": "#/components/schemas/StatusReason"
          },
          "alert": {
            "$ref": "#/components/schemas/Alert"
          },
          "dashboard": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MetricDescriptor"
            },
            "maxItems": 12,
            "minItems": 1
          },
          "recovery": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecoveryState"
              },
              {
                "type": "null"
              }
            ]
          },
          "timeLimitSeconds": {
            "type": "integer",
            "minimum": 60,
            "maximum": 14400
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "readyAt": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "endsAt": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "endedAt": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "hints": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "released": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Hint"
                },
                "maxItems": 20
              },
              "remaining": {
                "type": "integer",
                "minimum": 0,
                "maximum": 20
              },
              "nextAvailableAt": {
                "type": ["string", "null"],
                "format": "date-time"
              }
            },
            "required": ["released", "remaining", "nextAvailableAt"]
          },
          "assistance": {
            "$ref": "#/components/schemas/Assistance"
          },
          "debriefAvailable": {
            "type": "boolean"
          },
          "recording": {
            "$ref": "#/components/schemas/RecordingState"
          }
        },
        "required": [
          "id",
          "challenge",
          "attempt",
          "status",
          "statusReason",
          "alert",
          "dashboard",
          "recovery",
          "timeLimitSeconds",
          "createdAt",
          "readyAt",
          "endsAt",
          "endedAt",
          "hints",
          "assistance",
          "debriefAvailable",
          "recording"
        ]
      },
      "SessionResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "session": {
            "$ref": "#/components/schemas/SessionView"
          },
          "replayed": {
            "type": "boolean"
          }
        },
        "required": ["session", "replayed"]
      },
      "SessionSummary": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "challenge": {
            "$ref": "#/components/schemas/ChallengeRef"
          },
          "attempt": {
            "$ref": "#/components/schemas/Attempt"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "endedAt": {
            "type": ["string", "null"],
            "format": "date-time"
          }
        },
        "required": ["id", "challenge", "attempt", "status", "createdAt", "endedAt"]
      },
      "SessionPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionSummary"
            },
            "maxItems": 100
          },
          "nextCursor": {
            "type": ["string", "null"],
            "minLength": 1,
            "maxLength": 1000
          }
        },
        "required": ["items", "nextCursor"]
      },
...
      "RecordingState": {
        "description": "Recording is separate from the session outcome. An incomplete recording has no final numeric score.",
        "oneOf": [
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "status": {
                "enum": ["pending", "recording", "draining", "complete"]
              },
              "reason": {
                "type": "null"
              }
            },
            "required": ["status", "reason"]
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "status": {
                "const": "incomplete"
              },
              "reason": {
                "type": "string",
                "enum": ["task_lost", "recorder_lost", "drain_timeout", "limit_exceeded", "data_gap"]
              }
            },
            "required": ["status", "reason"]
          }
        ]
      }
    }
  }
}